chore: equalise V0 and V1 protocol definitions. (#115)
Today we define and publish V1 along side V0. This is a remnant from the V1 breaking change a few years back that we are no longer doing.

Unfortunately this means we now have different components using both V1 and V0 code packages. It's also easy for us to accidentally update one file and not the other - drift exists today. All of this is confusing.

I want to remove V1 from the protocol repo and move everyone over to V0. We'll reintroduce V1 when we are properly migrating. Step 1 is equalising both files.