feat: remove v1 models (#133)
The v1 protocol effort stalled, but we still have the v1 airbyte_protocol.yaml file in the repo. This is confusing, because it's unclear where updates to the protocol should be made. Discussed with @davinchia here that we should just remove the v1 file and fix forward any potential fallout.

Co-authored-by: Davin Chia <davinchia@gmail.com>